itest: move tests by their category
