+++ lntest/itest/lnd_misc_test.go
@@ -1323,209 +1323,6 @@ func testNodeSignVerify(net *lntest.NetworkHarness, t *harnessTest) {
 	closeChannelAndAssert(t, net, net.Alice, aliceBobCh, false)
 }
 
-// testSendUpdateDisableChannel ensures that a channel update with the disable
-// flag set is sent once a channel has been either unilaterally or cooperatively
-// closed.
-func testSendUpdateDisableChannel(net *lntest.NetworkHarness, t *harnessTest) {
-	ctxb := context.Background()
-
-	const (
-		chanAmt = 100000
-	)
-
-	// Open a channel between Alice and Bob and Alice and Carol. These will
-	// be closed later on in order to trigger channel update messages
-	// marking the channels as disabled.
-	chanPointAliceBob := openChannelAndAssert(
-		t, net, net.Alice, net.Bob,
-		lntest.OpenChannelParams{
-			Amt: chanAmt,
-		},
-	)
-
-	carol := net.NewNode(
-		t.t, "Carol", []string{
-			"--minbackoff=10s",
-			"--chan-enable-timeout=1.5s",
-			"--chan-disable-timeout=3s",
-			"--chan-status-sample-interval=.5s",
-		})
-	defer shutdownAndAssert(net, t, carol)
-
-	net.ConnectNodes(t.t, net.Alice, carol)
-	chanPointAliceCarol := openChannelAndAssert(
-		t, net, net.Alice, carol,
-		lntest.OpenChannelParams{
-			Amt: chanAmt,
-		},
-	)
-
-	// We create a new node Eve that has an inactive channel timeout of
-	// just 2 seconds (down from the default 20m). It will be used to test
-	// channel updates for channels going inactive.
-	eve := net.NewNode(
-		t.t, "Eve", []string{
-			"--minbackoff=10s",
-			"--chan-enable-timeout=1.5s",
-			"--chan-disable-timeout=3s",
-			"--chan-status-sample-interval=.5s",
-		})
-	defer shutdownAndAssert(net, t, eve)
-
-	// Give Eve some coins.
-	net.SendCoins(t.t, btcutil.SatoshiPerBitcoin, eve)
-
-	// Connect Eve to Carol and Bob, and open a channel to carol.
-	net.ConnectNodes(t.t, eve, carol)
-	net.ConnectNodes(t.t, eve, net.Bob)
-
-	chanPointEveCarol := openChannelAndAssert(
-		t, net, eve, carol,
-		lntest.OpenChannelParams{
-			Amt: chanAmt,
-		},
-	)
-
-	// Launch a node for Dave which will connect to Bob in order to receive
-	// graph updates from. This will ensure that the channel updates are
-	// propagated throughout the network.
-	dave := net.NewNode(t.t, "Dave", nil)
-	defer shutdownAndAssert(net, t, dave)
-
-	net.ConnectNodes(t.t, net.Bob, dave)
-
-	daveSub := subscribeGraphNotifications(ctxb, t, dave)
-	defer close(daveSub.quit)
-
-	// We should expect to see a channel update with the default routing
-	// policy, except that it should indicate the channel is disabled.
-	expectedPolicy := &lnrpc.RoutingPolicy{
-		FeeBaseMsat:      int64(chainreg.DefaultBitcoinBaseFeeMSat),
-		FeeRateMilliMsat: int64(chainreg.DefaultBitcoinFeeRate),
-		TimeLockDelta:    chainreg.DefaultBitcoinTimeLockDelta,
-		MinHtlc:          1000, // default value
-		MaxHtlcMsat:      calculateMaxHtlc(chanAmt),
-		Disabled:         true,
-	}
-
-	// Let Carol go offline. Since Eve has an inactive timeout of 2s, we
-	// expect her to send an update disabling the channel.
-	restartCarol, err := net.SuspendNode(carol)
-	if err != nil {
-		t.Fatalf("unable to suspend carol: %v", err)
-	}
-	waitForChannelUpdate(
-		t, daveSub,
-		[]expectedChanUpdate{
-			{eve.PubKeyStr, expectedPolicy, chanPointEveCarol},
-		},
-	)
-
-	// We restart Carol. Since the channel now becomes active again, Eve
-	// should send a ChannelUpdate setting the channel no longer disabled.
-	if err := restartCarol(); err != nil {
-		t.Fatalf("unable to restart carol: %v", err)
-	}
-
-	expectedPolicy.Disabled = false
-	waitForChannelUpdate(
-		t, daveSub,
-		[]expectedChanUpdate{
-			{eve.PubKeyStr, expectedPolicy, chanPointEveCarol},
-		},
-	)
-
-	// Now we'll test a long disconnection. Disconnect Carol and Eve and
-	// ensure they both detect each other as disabled. Their min backoffs
-	// are high enough to not interfere with disabling logic.
-	if err := net.DisconnectNodes(carol, eve); err != nil {
-		t.Fatalf("unable to disconnect Carol from Eve: %v", err)
-	}
-
-	// Wait for a disable from both Carol and Eve to come through.
-	expectedPolicy.Disabled = true
-	waitForChannelUpdate(
-		t, daveSub,
-		[]expectedChanUpdate{
-			{eve.PubKeyStr, expectedPolicy, chanPointEveCarol},
-			{carol.PubKeyStr, expectedPolicy, chanPointEveCarol},
-		},
-	)
-
-	// Reconnect Carol and Eve, this should cause them to reenable the
-	// channel from both ends after a short delay.
-	net.EnsureConnected(t.t, carol, eve)
-
-	expectedPolicy.Disabled = false
-	waitForChannelUpdate(
-		t, daveSub,
-		[]expectedChanUpdate{
-			{eve.PubKeyStr, expectedPolicy, chanPointEveCarol},
-			{carol.PubKeyStr, expectedPolicy, chanPointEveCarol},
-		},
-	)
-
-	// Now we'll test a short disconnection. Disconnect Carol and Eve, then
-	// reconnect them after one second so that their scheduled disables are
-	// aborted. One second is twice the status sample interval, so this
-	// should allow for the disconnect to be detected, but still leave time
-	// to cancel the announcement before the 3 second inactive timeout is
-	// hit.
-	if err := net.DisconnectNodes(carol, eve); err != nil {
-		t.Fatalf("unable to disconnect Carol from Eve: %v", err)
-	}
-	time.Sleep(time.Second)
-	net.EnsureConnected(t.t, eve, carol)
-
-	// Since the disable should have been canceled by both Carol and Eve, we
-	// expect no channel updates to appear on the network.
-	assertNoChannelUpdates(t, daveSub, 4*time.Second)
-
-	// Close Alice's channels with Bob and Carol cooperatively and
-	// unilaterally respectively.
-	_, _, err = net.CloseChannel(net.Alice, chanPointAliceBob, false)
-	if err != nil {
-		t.Fatalf("unable to close channel: %v", err)
-	}
-
-	_, _, err = net.CloseChannel(net.Alice, chanPointAliceCarol, true)
-	if err != nil {
-		t.Fatalf("unable to close channel: %v", err)
-	}
-
-	// Now that the channel close processes have been started, we should
-	// receive an update marking each as disabled.
-	expectedPolicy.Disabled = true
-	waitForChannelUpdate(
-		t, daveSub,
-		[]expectedChanUpdate{
-			{net.Alice.PubKeyStr, expectedPolicy, chanPointAliceBob},
-			{net.Alice.PubKeyStr, expectedPolicy, chanPointAliceCarol},
-		},
-	)
-
-	// Finally, close the channels by mining the closing transactions.
-	mineBlocks(t, net, 1, 2)
-
-	// Also do this check for Eve's channel with Carol.
-	_, _, err = net.CloseChannel(eve, chanPointEveCarol, false)
-	if err != nil {
-		t.Fatalf("unable to close channel: %v", err)
-	}
-
-	waitForChannelUpdate(
-		t, daveSub,
-		[]expectedChanUpdate{
-			{eve.PubKeyStr, expectedPolicy, chanPointEveCarol},
-		},
-	)
-	mineBlocks(t, net, 1, 1)
-
-	// And finally, clean up the force closed channel by mining the
-	// sweeping transaction.
-	cleanupForceClose(t, net, net.Alice, chanPointAliceCarol)
-}
-
 // testAbandonChannel abandones a channel and asserts that it is no
 // longer open and not in one of the pending closure states. It also
 // verifies that the abandoned channel is reported as closed with close
